refactor, docs
