+++ src/Perpetuum/Services/EventServices/EventMessages/EventMessage.cs
@@ -1,0 +1,9 @@
+﻿namespace Perpetuum.Services.EventServices.EventMessages
+{
+    /// <summary>
+    /// A specification Interface for EventMessages
+    /// </summary>
+    public interface EventMessage
+    {
+    }
+}
