+++ src/Perpetuum/Zones/NpcSystem/Npc.cs
@@ -16,7 +16,6 @@ using Perpetuum.Modules.Weapons;
 using Perpetuum.PathFinders;
 using Perpetuum.Players;
 using Perpetuum.Services.EventServices;
-using Perpetuum.Services.EventServices.EventProcessors;
 using Perpetuum.Services.Looting;
 using Perpetuum.Services.MissionEngine;
 using Perpetuum.Services.MissionEngine.MissionTargets;
@@ -25,7 +24,6 @@ using Perpetuum.Timers;
 using Perpetuum.Units;
 using Perpetuum.Zones.DamageProcessors;
 using Perpetuum.Zones.Eggs;
-using Perpetuum.Zones.Intrusion;
 using Perpetuum.Zones.Locking;
 using Perpetuum.Zones.Locking.Locks;
 using Perpetuum.Zones.Movements;
